csi: update of terminal plugin allocs should clean themselves up
When an allocation that implements a CSI plugin becomes terminal the
client fingerprint can't tell if the plugin is unhealthy intentionally
(for the case of updates or job stop). Allocations that are
server-terminal should delete themselves from the plugin and trigger a
plugin self-GC, the same as an unused node.